Adds sending mail on new top level comment.
